Roll-up changes for point release 3.7 for next pathfinder

+++ lsstsal/scripts/updatevisitsim.tcl
@@ -11,8 +11,23 @@ source $SAL_DIR/sal_version.tcl
 set DEST /data/staging/ts_visit_simulator
 
 exec mkdir -p $DEST/test/tcs
+exec mkdir -p $DEST/test/bin
 exec mkdir -p $DEST/test/lib
 exec mkdir -p $DEST/test/include
+
+echo "Modifying hexapod positionSet commander"
+cd $SAL_WORK_DIR/hexapod/cpp/src
+exec cp $SAL_DIR/sacpp_hexapod_positionSet_commander.cpp .
+exec make -f /tmp/Makefile.sacpp_hexapod_testcommands
+
+cd $SAL_WORK_DIR/tcs/tcs/demo
+exec make
+if { [file exists tpkDemo.so] == 0 } {
+   puts stdout "************FAILED TO BUILD tpkDemo.so****************"
+   exit
+}
+
+exec mv tpkDemo.so ../bin/.
 
 puts stdout "Copying tcs simulator"
 exec cp -rv $SAL_WORK_DIR/tcs/tcs $DEST/test/tcs/.
@@ -62,7 +77,7 @@ set all [glob $SAL_WORK_DIR/*/labview]
 foreach i $all {
    exec cp -rv $i $DEST/test/.
 }
-}
+
 
 puts stdout "Updating libSALPY"
 set all [glob $SAL_DIR/../lib/SALPY_*.so]
@@ -129,13 +144,41 @@ foreach subsys $SYSDIC(systems) {
    }
   }
 }
+}
 
+puts stdout "Updating commanders and controllers"
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_initImage_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_setFilter_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_takeImages_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/dome/cpp/src/sacpp_dome_Move_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/dome/cpp/src/sacpp_dome_Crawl_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/hexapod/cpp/src/sacpp_hexapod_positionSet_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/m2ms/cpp/src/sacpp_m2ms_ApplyBendingMode_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_startIntegration_send $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_startReadout_send $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/rotator/cpp/src/sacpp_rotator_track_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/MTMount/cpp/src/sacpp_MTMount_trackTarget_commander $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/hexapod/cpp/src/sacpp_hexapod_move_commander $DEST/test/bin/.
+
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_initImage_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_setFilter_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_takeImages_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/dome/cpp/src/sacpp_dome_Move_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/dome/cpp/src/sacpp_dome_Crawl_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/hexapod/cpp/src/sacpp_hexapod_positionSet_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/m2ms/cpp/src/sacpp_m2ms_ApplyBendingMode_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_startIntegration_log $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/camera/cpp/src/sacpp_camera_startReadout_log $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/rotator/cpp/src/sacpp_rotator_track_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/MTMount/cpp/src/sacpp_MTMount_trackTarget_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/hexapod/cpp/src/sacpp_hexapod_move_controller $DEST/test/bin/.
+exec cp $SAL_WORK_DIR/hexapod/cpp/src/sacpp_hexapod_move_controller $DEST/test/bin/.
 
 puts stdout "Updating EFD clients"
 source $env(SAL_DIR)/gengenericefd.tcl
 
 updateefdschema
-foreach subsys $SYSDIC(systems) {
+foreach subsys "camera dome hexapod MTMount m2ms rotator" {
   catch {
    set all [glob $SAL_WORK_DIR/[set subsys]/cpp/src/sacpp_*_efdwriter]
    foreach i $all {
